requirements.txt
flask==3.0.3
gunicorn==22.0.0

# MediaPipe Gesture Recognizer
mediapipe==0.10.32

# Image decoding
pillow==12.1.1
numpy==1.26.4
